Refactor config migrations from functions into a proper interface for re-usability and management of KV values
